Switched perspective matrix from GLM to my Mat4. Added a general purpose rotate function to Mat4 to make the calling process similar to Quats. Changed MemoryCounter to make it more precise. Added a backup memory measurement if the MemoryCounter can not detect any change. Created basic benchmark process. Added controls to change the parameters. Started implementing the changing of angles and axis.

+++ GCP Assignment 1/misc/MemoryCounter.h
@@ -3,5 +3,5 @@
 class MemoryCounter
 {
 public:
-	static float getMemoryUsage();
+	static long long getMemoryUsage();
 };
